Move `UTxO` related modules to `Pure.UTxO.*`
