+++ lib/customer-deposit-wallet/src/Cardano/Wallet/Deposit/Pure.hs
@@ -38,7 +38,7 @@ module Cardano.Wallet.Deposit.Pure
 
 import Prelude
 
-import Cardano.Wallet.Deposit.Pure.UTxOHistory
+import Cardano.Wallet.Deposit.Pure.UTxO.UTxOHistory
     ( UTxOHistory
     )
 import Cardano.Wallet.Deposit.Read
@@ -60,10 +60,10 @@ import Numeric.Natural
     ( Natural
     )
 
-import qualified Cardano.Wallet.Deposit.Pure.Balance as Balance
 import qualified Cardano.Wallet.Deposit.Pure.Submissions as Sbm
-import qualified Cardano.Wallet.Deposit.Pure.UTxO as UTxO
-import qualified Cardano.Wallet.Deposit.Pure.UTxOHistory as UTxOHistory
+import qualified Cardano.Wallet.Deposit.Pure.UTxO.Balance as Balance
+import qualified Cardano.Wallet.Deposit.Pure.UTxO.UTxO as UTxO
+import qualified Cardano.Wallet.Deposit.Pure.UTxO.UTxOHistory as UTxOHistory
 import qualified Cardano.Wallet.Deposit.Read as Read
 import qualified Cardano.Wallet.Deposit.Write as Write
 import qualified Data.Delta as Delta
