+++ lib/customer-deposit-wallet/src/Cardano/Wallet/Deposit/Pure/UTxO/UTxO.hs
@@ -1,4 +1,4 @@
-module Cardano.Wallet.Deposit.Pure.UTxO
+module Cardano.Wallet.Deposit.Pure.UTxO.UTxO
     ( UTxO (..)
     , balance
     , excluding
